% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{Format a p value for presentation in a R Markdown or Quarto document}
\usage{
format_p(p_value, digits = 3)
}
\arguments{
\item{p_value}{The p value to format}

\item{digits}{The number of digits to round the p value to}
}
\value{
A string with the formatted p value
}
\description{
Format a p value for presentation in a R Markdown or Quarto document
}
\examples{
p <- 0.00445463
format_p(p)
}
