% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cor.R
\name{report_cor}
\alias{report_cor}
\title{Report correlation result in text}
\usage{
report_cor(cor_test, effect = 1, digits = 3)
}
\arguments{
\item{cor_test}{an rstatix::cor_test object}

\item{effect}{the names or row number of the Effect to report}

\item{digits}{the number of digits to round the p value to}
}
\value{
a string with the formatted correlation result for use inline in an R Markdown or Quarto document
}
\description{
Report correlation result in text
}
\examples{
library(rstatix)

results <- mtcars |> cor_test(mpg, wt)
report_cor(results)

results2 <- mtcars |>
 group_by(cyl) |>
 cor_test(mpg, wt, method = "spearman")

report_cor(results2, effect = 2)

results3 <- mtcars \%>\% cor_test(
    vars = "mpg",
    vars2 = c("disp", "hp", "drat"),
    method = "kendall"
  )

 report_cor(results3, effect = c("mpg", "disp"))
}
