% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mean_var.R
\name{report_mean_sem}
\alias{report_mean_sem}
\title{Report the mean and standard error of the mean in text}
\usage{
report_mean_sem(.data, value_var, group_var = NULL, effect = 1, digits = 3)
}
\arguments{
\item{.data}{A data frame}

\item{value_var}{The name of the variable to report}

\item{group_var}{The name of the grouping variable}

\item{effect}{The name or row number of the effect to report}

\item{digits}{The number of digits to round the p value to}
}
\value{
A string with the mean and standard error of the variable in the format: "mean ± sem"
}
\description{
Report the mean and standard error of the mean in text
}
\examples{
mtcars |> report_mean_sem(value_var = mpg)
mtcars |> report_mean_sem(mpg, cyl, effect = "6")
}
