% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_t.R
\name{report_t}
\alias{report_t}
\title{Report a t test result in text}
\usage{
report_t(
  ttest,
  effect = 1,
  digits = 3,
  cohensd = NULL,
  cohens_magnitude = FALSE
)
}
\arguments{
\item{ttest}{an rstatix::t_test object}

\item{effect}{The names or row number of the effect to report}

\item{digits}{The number of digits to round the p value to}

\item{cohensd}{An rstatix::cohens_d object to include the effect size in the report}

\item{cohens_magnitude}{A logical indicating if the effect size magnitude should be included in the report}
}
\value{
A string with the formatted t test result for use inline in an R Markdown or Quarto document
}
\description{
Report a t test result in text
}
\examples{
library(rstatix)

results <- t_test(mtcars, mpg ~ am)

report_t(results)

results2 <- t_test(mtcars, mpg ~ cyl)

cohensd2 <- cohens_d(mtcars, mpg ~ cyl)

report_t(results2, cohensd = cohensd2, cohens_magnitude = TRUE)

report_t(results2, effect = c("4", "6"))
}
