% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_pc.R
\name{report_pc}
\alias{report_pc}
\title{Report a pairwise comparison result in text}
\usage{
report_pc(pairwise_comparison, effect = 1, digits = 3)
}
\arguments{
\item{pairwise_comparison}{an rstatix::emmeans_test object}

\item{effect}{The names or row number of the effect to report}

\item{digits}{The number of digits to round the p value to}
}
\value{
A string with the formatted pairwise comparison result for use inline in an R Markdown or Quarto document
}
\description{
Report a pairwise comparison result in text
}
\examples{
library(rstatix)

results <- emmeans_test(mtcars, mpg ~ am)

report_pc(results)

results2 <- emmeans_test(mtcars, mpg ~ cyl, p.adjust.method = "none")

report_pc(results2)

report_pc(results2, effect = c("4", "6"))
}
